derive_tools: refactor from
