+++ module/core/former_meta/src/derive_former/field_attrs.rs
@@ -3,7 +3,6 @@
 use super::*;
 use macro_tools::
 {
-  attr,
   Result,
   AttributeComponent,
   AttributePropertyComponent,
