+++ module/core/former_meta/src/derive_former/struct_attrs.rs
@@ -6,7 +6,6 @@ use super::*;
 
 use macro_tools::
 {
-  attr,
   Result,
   AttributeComponent,
   AttributePropertyComponent,
